services/paper-chat/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
supabase==2.8.1
sentence-transformers==3.1.1
torch==2.4.1
numpy==1.26.4
PyMuPDF==1.24.10
pdfplumber==0.11.4
langdetect==1.0.9
langchain==0.3.0
langchain-text-splitters==0.3.0
tiktoken==0.7.0
python-dotenv==1.0.1
